third_party/WebKit/LayoutTests/media/video-controls-muted-video-can-unmute.html

<video controls></video>
<p>Test that muted video has an unmute button.<p>
<script src=media-file.js></script>
<script src=video-test.js></script>
<script src=media-controls.js></script>
<script>
    run("video.muted = true");
    video.src = findMediaFile("video", "content/test");
    run("video.load()");
    waitForEvent("canplaythrough", function () {
        testExpected("video.muted", true);
        muteButton = mediaControlsButton(video, "mute-button");
        // Make sure that the mute button is visible.
        testExpected("getComputedStyle(muteButton).display", "none", '!=');
        endTest();
    });
</script>
